Relocate issue and redeem to the GiftCardService (#2)
* Change exception handling for Guzzle clients
Instead of throwing an exception on 4XX and 5XX errors, the responses are now returned so they can be handled by the services themselves.

* Add cleanCode method on GiftCard resource
The logic of cleaning user input has been extracted to a static method on the GiftCard resource. This allows the reuse of the same logic to keep consistency.

* Move issue and redeem methods to the GiftCardService
The issue and redeem methods apply to a GiftCard and not to a Transaction. This also removes the requirement of retrieving a GiftCard before a new issue or redeem action can be created.