Remove guzzle dependency (#5)
* Update GiftyGuzzleHttpClient.php

The current Guzzle implementation can cause issues when a site also has the previous Guzzle 6 installed because of a renamed constant.

* Update constant retrieval so undefined constants are handled

* Replaced the Guzzle client by a custom Http client
The Guzzle client dependency caused issues for implementations that do not support package managers like WordPress. By removing all third party dependencies the risk of interference is removed.
PSR compatability is maintained so that it is still possible to manually pass a Guzzle client when constructing the Gifty Client.

* Replaced the Guzzle client by a custom Http client
The Guzzle client dependency caused issues for implementations that do not support package managers like WordPress. By removing all third party dependencies the risk of interference is removed.
PSR compatability is maintained so that it is still possible to manually pass a Guzzle client when constructing the Gifty Client.