Allow setting custom HTTP headers (#20)
It can be useful to pass different headers to the Gifty API. E.g. to pass the preferred language with the Accept-Language header.

This commit allows to pass custom headers via the $options parameter when initiating the GiftyClient. To do so, it is possible to pass 'api_headers' in the options array. See the example below.

$gifty = new \Gifty\Client\GiftyClient('eyJ0eXAi....', ['api_headers' => [
    'Accept-Language' => 'nl'
]]);